+++ packages/block-library/src/query/utils.js
@@ -10,7 +10,12 @@ import { useSelect } from '@wordpress/data';
 import { useMemo } from '@wordpress/element';
 import { store as coreStore } from '@wordpress/core-data';
 import { decodeEntities } from '@wordpress/html-entities';
-import { cloneBlock } from '@wordpress/blocks';
+import { cloneBlock, store as blocksStore } from '@wordpress/blocks';
+
+/**
+ * Internal dependencies
+ */
+import { name as queryLoopName } from './block.json';
 
 /**
  * @typedef IHasNameAndId
@@ -126,6 +131,48 @@ export const useTaxonomies = ( postType ) => {
 	);
 	return taxonomies;
 };
+
+/**
+ * Hook that returns whether a specific post type is hierarchical.
+ *
+ * @param {string} postType The post type to check.
+ * @return {boolean} Whether a specific post type is hierarchical.
+ */
+export function useIsPostTypeHierarchical( postType ) {
+	return useSelect(
+		( select ) => {
+			const type = select( coreStore ).getPostType( postType );
+			return type?.viewable && type?.hierarchical;
+		},
+		[ postType ]
+	);
+}
+
+/**
+ * Hook that returns the query properties' names defined by the active
+ * block variation, to determine which block's filters to show.
+ *
+ * @param {Object} attributes Block attributes.
+ * @return {string[]} An array of the query attributes.
+ */
+export function useAllowedControls( attributes ) {
+	return useSelect(
+		( select ) =>
+			select( blocksStore ).getActiveBlockVariation(
+				queryLoopName,
+				attributes
+			)?.allowControls,
+
+		[ attributes ]
+	);
+}
+export function isControllAllowed( allowedControls, key ) {
+	// Every controls is allowed if the list is not defined.
+	if ( ! allowedControls ) {
+		return true;
+	}
+	return allowedControls.includes( key );
+}
 
 /**
  * Clones a pattern's blocks and then recurses over that list of blocks,
